requirements.txt
pyTelegramBotAPI==4.26.0
supabase==2.13.0
openai==1.65.2
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.38.0
pydantic==2.10.6
